Creating gof.cpp
This is the C++ code that I use to calculate these statistics. There are some references to outside classes, but these are mostly self-explanatory. `Table` is a container of structs, whose values include `pi`, the regressed probability; `dpi`, the gradient of the regression function at the fit point; `correct`, a boolean corresponding to Y above; and `W`, a sample weight which is identically equal to 1 in all these tests. `selfCalibrated` is true in all these tests (it would be false in a validation setting).